refactor: split tools/common.go into focused files and clean up
Split the 791-line tools/common.go into 7 focused files:
- helpers.go: shared utilities (toolErr, parseKey, waitDOMStable, resolveSnapshotElement)
- navigation.go: navigate, go back, go forward, reload
- input.go: press key, file upload
- media.go: screenshot, PDF
- browser.go: evaluate, close browser, set headers, resize, handle dialog
- tabs.go: tab new/list/select/close
- debug.go: wait for, console messages, network requests

Additional fixes:
- Deduplicate navigation handlers via simplePageAction helper (#56)
- Deduplicate snapshot element resolution via resolveSnapshotElement helper (#56)
- Add waitDOMStable helper that logs errors at debug level (#57)
- Remove unused screenshot params (selector, width, height) (#58)
- Use NavigationToolKey constant instead of hardcoded string (#58)
- Rename ReLoad/ReLoadHandler to Reload/ReloadHandler (#58)

Fixes #55, fixes #56, fixes #57, fixes #58
